Move LogEntry to separate file
